работа с мультипоточностью и аннотациями
