+++ src/test/ru/kampaii/examples/repositories/UserServiceCommonTest.java
@@ -1,10 +1,13 @@
 package ru.kampaii.examples.repositories;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
 import ru.kampaii.examples.domain.entities.AccountsEntity;
 import ru.kampaii.examples.domain.entities.UsersEntity;
-import ru.kampaii.examples.repositories.servises.UserService;
 import ru.kampaii.examples.repositories.servises.UserServiceCommon;
 
 import java.sql.SQLException;
@@ -12,20 +15,22 @@ import java.sql.SQLException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 
+@ExtendWith(MockitoExtension.class)
 public class UserServiceCommonTest {
+    @Mock
     private Repository<UsersEntity, Integer> usersRepository;
+    @Mock
     private Repository<AccountsEntity, Integer> accountsRepository;
+    @InjectMocks
+    private UserServiceCommon service;
 
     @Test
     void userServiceUnitTest() throws SQLException {
-        usersRepository = Mockito.mock(UsersRepositoryImpl.class);
-        accountsRepository = Mockito.mock(AccountsRepositoryImpl.class);
-        UserService service = new UserServiceCommon(usersRepository, accountsRepository);
         String name = "TEST";
         Mockito.when(usersRepository.create(any())).thenReturn(new UsersEntity(1, name, 0F));
-        Mockito.when(accountsRepository.create(any())).thenReturn(new AccountsEntity(1, 0F, 1, 1));
         UsersEntity entity = service.createUser(name, 5);
         assertEquals(name, entity.getName());
     }
 
 }
+
