+++ src/test/ru/kampaii/examples/repositories/UserServicesTest.java
@@ -8,6 +8,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import ru.kampaii.examples.config.DatabaseConnectorProvider;
 import ru.kampaii.examples.domain.entities.AccountsEntity;
 import ru.kampaii.examples.domain.entities.UsersEntity;
+import ru.kampaii.examples.repositories.id.generators.IdGenerator;
 import ru.kampaii.examples.repositories.id.generators.PooledIdGeneratorImpl;
 import ru.kampaii.examples.repositories.servises.UserService;
 import ru.kampaii.examples.repositories.servises.UserServiceCommon;
@@ -16,18 +17,27 @@ import ru.kampaii.examples.repositories.servises.UserServiceTransactionalBatch;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(MockitoExtension.class)
-public class UserServicesTest {
+class UserServicesTest {
     private Connection connection;
     private Repository<UsersEntity, Integer> usersRepository;
     private Repository<AccountsEntity, Integer> accountsRepository;
+    private IdGenerator<Integer> usersIdGenerator;
+    private IdGenerator<Integer> accountsIdGenerator;
 
     @BeforeEach
     void setUp() throws SQLException {
         this.connection = DatabaseConnectorProvider.connect();
+        this.usersRepository = new UsersRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "users", "id", 1, 1000));
+        this.accountsRepository = new AccountsRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "accounts", "number", 1, 1000));
     }
 
     @AfterEach
@@ -38,24 +48,18 @@ public class UserServicesTest {
 
     @Test
     void userServiceTest() throws SQLException {
-        usersRepository = new UsersRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "users", "id", 1, 1000));
-        accountsRepository = new AccountsRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "accounts", "number", 1, 1000));
         UserService service = new UserServiceCommon(usersRepository, accountsRepository);
         executeIntegrationTest(service);
     }
 
     @Test
     void userServiceTransTest() throws SQLException {
-        usersRepository = new UsersRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "users", "id", 1, 1000));
-        accountsRepository = new AccountsRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "accounts", "number", 1, 1000));
         UserService service = new UserServiceTransactional(usersRepository, accountsRepository, connection);
         executeIntegrationTest(service);
     }
 
     @Test
     void userServiceTransBatchTest() throws SQLException {
-        usersRepository = new UsersRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "users", "id", 1, 1000));
-        accountsRepository = new AccountsRepositoryPreparedImpl(connection, new PooledIdGeneratorImpl(connection, "accounts", "number", 1, 1000));
         UserService service = new UserServiceTransactionalBatch(usersRepository, accountsRepository, connection);
         executeIntegrationTest(service);
     }
@@ -68,6 +72,35 @@ public class UserServicesTest {
         assertEquals(name, entity.getName());
         assertEquals(5, accountsRepository.count() - firstAccountsCount);
         assertEquals(1, usersRepository.count() - firstUsersCount);
+    }
+
+    @Test
+    void testWithMultiThreads() throws SQLException, InterruptedException {
+        int firstAccountsCount = accountsRepository.count();
+        int firstUsersCount = usersRepository.count();
+        usersIdGenerator = new PooledIdGeneratorImpl(DatabaseConnectorProvider.connect(), "users", "id", 1, 10000);
+        accountsIdGenerator = new PooledIdGeneratorImpl(DatabaseConnectorProvider.connect(), "accounts", "number", 1, 10000);
+        ExecutorService executor = Executors.newFixedThreadPool(5);
+        List<Runnable> listForRunnable = new ArrayList<>();
+        for (int i = 0; i < 5; i++) {
+            listForRunnable.add(this::executeTestForSingleThread);
+        }
+        listForRunnable.forEach(executor::submit);
+        executor.awaitTermination(5, SECONDS);
+        assertEquals(5000, accountsRepository.count() - firstAccountsCount);
+        assertEquals(1000, usersRepository.count() - firstUsersCount);
+    }
+
+    public void executeTestForSingleThread() {
+        try {
+            Connection connectionForThread = DatabaseConnectorProvider.connect();
+            UserService userService = new UserServiceTransactionalBatch(new UsersRepositoryImpl(connectionForThread, usersIdGenerator), new AccountsRepositoryImpl(connectionForThread, accountsIdGenerator), connectionForThread);
+            for (int i = 0; i < 200; i++) {
+                userService.createUser(String.valueOf(i), 5);
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
     }
 
 }
